requirements.txt
pandas==2.2.1
Pillow==10.2.0
python-dotenv==1.0.1
Requests==2.31.0
selenium==4.18.1
